No more trait, DTO is init in reader constructor
