+++ Tests/Reader/XmlReaderTest.php
@@ -24,6 +24,7 @@ class XmlReaderTest extends TestCase
         // init reader
         new XmlReader(
             __DIR__.'/../fixtures/xml/books_with_wrong_path.xml',
+            null,
             [XmlReader::CONTEXT_XPATH => 'shop/catalog']
         );
     }
@@ -33,11 +34,12 @@ class XmlReaderTest extends TestCase
         // init reader
         $reader = new XmlReader(
             __DIR__.'/../fixtures/xml/books_with_xpath.xml',
+            null,
             [XmlReader::CONTEXT_XPATH => 'shop/catalog']
         );
 
         // test denormalization
-        $this->assertTrue($reader->isDenormalizable());
+        $this->assertFalse($reader->isDenormalizable());
 
         // test count
         $this->assertEquals(2, count($reader));
@@ -87,10 +89,13 @@ class XmlReaderTest extends TestCase
     public function testReadXmlWithoutXpath()
     {
         // init reader
-        $reader = new XmlReader(__DIR__.'/../fixtures/xml/books_without_xpath.xml');
+        $reader = new XmlReader(
+            __DIR__.'/../fixtures/xml/books_without_xpath.xml',
+            null
+        );
 
         // test denormalization
-        $this->assertTrue($reader->isDenormalizable());
+        $this->assertFalse($reader->isDenormalizable());
 
         // test count
         $this->assertEquals(2, count($reader));
@@ -145,6 +150,7 @@ class XmlReaderTest extends TestCase
         // init reader
         new XmlReader(
             __DIR__.'/../fixtures/xml/books_with_xpath.xml',
+            null,
             [XmlReader::CONTEXT_XPATH => 'foo']
         );
     }
@@ -157,6 +163,7 @@ class XmlReaderTest extends TestCase
         // init reader
         new XmlReader(
             __DIR__.'/../fixtures/xml/books_with_xpath.xml',
+            null,
             [XmlReader::CONTEXT_XPATH => 'shop/foo']
         );
     }
